Refactor some code to avoid duplicated code blocks
